LayoutTests/fast/scrolling/mac/wheel-event-listener-region-tiled-layer-switch-expected.txt


Test that scroll over box does not scroll page
PASS windowScrollEventCount is 0

Test that scroll over box does not scroll page after switch to tiled layer
PASS windowScrollEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
